requirements.txt
pywikibot 
datasets 
mwparserfromhell
python-dotenv
